refactor: 다음 라운드로 넘어가는 버튼 컴포넌트 분리 #50
