+++ frontend/src/pages/RoundResultPage/RoundResultPage.tsx
@@ -1,35 +1,14 @@
-import { useNavigate } from 'react-router-dom';
-
-import { moveNextRound } from '@/apis/balanceContent';
-import Button from '@/components/common/Button/Button';
+import NextRoundButton from '@/components/common/NextRoundButton/NextRoundButton';
 import Content from '@/components/layout/Content/Content';
 import RoundResultTab from '@/components/RoundResultTab/RoundResultTab';
 import TopicContainer from '@/components/TopicContainer/TopicContainer';
-import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
 
 const RoundResultPage = () => {
-  const navigate = useNavigate();
-  const { balanceContent } = useBalanceContentQuery();
-
-  const isLastRound = balanceContent?.currentRound === balanceContent?.totalRound;
-
-  const goToGameResult = () => {
-    navigate('/game/result');
-  };
-
-  const goToNextRound = async () => {
-    await moveNextRound();
-    navigate('/game');
-  };
-
   return (
     <Content>
       <TopicContainer />
       <RoundResultTab />
-      <Button
-        text={isLastRound ? '결과 확인' : '다음'}
-        onClick={isLastRound ? goToGameResult : goToNextRound}
-      />
+      <NextRoundButton />
     </Content>
   );
 };
